Create core.magics.namespace according to new API.
