+++ IPython/core/interactiveshell.py
@@ -2006,7 +2006,7 @@ class InteractiveShell(SingletonConfigurable):
         self.define_magic = self.magics_manager.define_magic
 
         self.register_magics(m.BasicMagics, m.CodeMagics, m.ConfigMagics,
-             mf.ExecutionMagics, mf.NamespaceMagics, mf.AutoMagics,
+             mf.ExecutionMagics, m.NamespaceMagics, mf.AutoMagics,
              mf.OSMagics, mf.LoggingMagics, mf.ExtensionsMagics,
              mf.PylabMagics, m.HistoryMagics, mf.DeprecatedMagics)
