+++ IPython/core/magics/__init__.py
@@ -17,6 +17,7 @@ from .basic import BasicMagics
 from .code import CodeMagics, MacroToEdit
 from .config import ConfigMagics
 from .history import HistoryMagics
+from .namespace import NamespaceMagics
 
 #-----------------------------------------------------------------------------
 # Magic implementation classes
